Exception handling - delete

+++ src/main/java/com/projetowebservices/course/services/UserService.java
@@ -4,12 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.projetowebservices.course.entities.User;
 import com.projetowebservices.course.repositories.UserRepository;
+import com.projetowebservices.course.resources.exceptions.DatabaseException;
 import com.projetowebservices.course.services.exceptions.ResourceNotFoundException;
-
+ 
 @Service
 public class UserService {
 
@@ -34,7 +37,31 @@ public class UserService {
 	
 	public void delete(Long id){
 		
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+		
+			throw new ResourceNotFoundException(id);
+		}
+		
+//		try {
+//			
+//			repository.deleteById(id);
+//				
+//		} catch (ResourceNotFoundException e) {
+//		
+//			throw new ResourceNotFoundException(id);
+//		}
+		
+		try {
+			repository.deleteById(id);
+
+		} catch (EmptyResultDataAccessException e) {
+
+			throw new ResourceNotFoundException(id);
+			
+		} catch (DataIntegrityViolationException e) {
+
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj){
